@@ -8,7 +8,7 @@
     "authors": [
         {
             "name": "Fusonic",
-            "homepage": "http://www.fusonic.net"
+            "homepage": "https://www.fusonic.net"
         }
     ],
     "autoload": {
@@ -22,13 +22,21 @@
         }
     },
     "require": {
-        "php": "^5.4|^7.0",
-        "symfony/dom-crawler": "^2.5",
-        "symfony/css-selector": "^2.5",
-        "fusonic/linq": "~1.0",
-        "guzzlehttp/guzzle": "^4.0"
+        "php": "^7.4",
+        "ext-dom": "*",
+        "symfony/dom-crawler": "^5.0",
+        "symfony/css-selector": "^5.0",
+        "fusonic/linq": "^1.0",
+        "psr/http-client": "^1.0",
+        "psr/http-factory": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8|^5.7"
+        "phpunit/phpunit": "^9.0",
+        "symfony/http-client": "^5.0",
+        "nyholm/psr7": "^1.2"
+    },
+    "suggest": {
+        "symfony/http-client": "^5.0",
+        "nyholm/psr7": "^1.2"
     }
 }
